refactor: rename claude.getDailyUsageData to claude.getDailyUsage
Renames the component identifier, Go types, payload type, and the
matching frontend mapper/file for consistency.
