+++ web_src/src/pages/app/mappers/claude/index.ts
@@ -1,14 +1,14 @@
 import type { ComponentBaseMapper, EventStateRegistry, TriggerRenderer } from "../types";
 import { baseMapper } from "./base";
 import { runCodeAgentMapper } from "./run_code_agent";
-import { getDailyUsageDataMapper } from "./get_daily_usage_data";
+import { getDailyUsageMapper } from "./get_daily_usage";
 import { buildActionStateRegistry } from "../utils";
 
 export const componentMappers: Record<string, ComponentBaseMapper> = {
   textPrompt: baseMapper,
   runAgent: baseMapper,
   runCodeAgent: runCodeAgentMapper,
-  getDailyUsageData: getDailyUsageDataMapper,
+  getDailyUsage: getDailyUsageMapper,
 };
 
 export const triggerRenderers: Record<string, TriggerRenderer> = {};
@@ -17,5 +17,5 @@ export const eventStateRegistry: Record<string, EventStateRegistry> = {
   textPrompt: buildActionStateRegistry("completed"),
   runAgent: buildActionStateRegistry("completed"),
   runCodeAgent: buildActionStateRegistry("completed"),
-  getDailyUsageData: buildActionStateRegistry("completed"),
+  getDailyUsage: buildActionStateRegistry("completed"),
 };
